Add support for SENSEGRP (sense groups)
Create a SenseGroup class, and have both it and
Sense keep a list of "sensestuff", the top-level
children within them that are senses, supplements,
and notes.

This allows us to recursively run through that list
for display purposes, finally getting the notes
where they're supposed to be.

Additionally, Entry::Sense objects now have a possibly-nil 'sensegrp_number'
that can be used for diplay, if needed.